+++ src/app/setting/toDeveloper/page.tsx
@@ -1,9 +1,9 @@
 "use client";
 import { theme } from "@/app/styles/theme";
 import SettingHeader from "@/components/settingHeader";
+import SettingInput from "@/components/settingInput";
 import { useState } from "react";
 import { BiMessageSquareEdit } from "react-icons/bi";
-import { FaCheckCircle, FaPlusCircle } from "react-icons/fa";
 import styled from "styled-components";
 
 const Wrapper = styled.div`
@@ -41,50 +41,12 @@ const Highlight = styled.div`
   font-weight: 800;
   line-height: normal;
 `;
-const InputContainer = styled.div`
-  width: 95%;
-  display: flex;
-  flex-direction: row;
-  margin: 2.13rem 0;
-  position: relative;
-`;
-const Icon = styled.div`
-  width: 1.25rem;
-  height: 1.25rem;
-  position: relative;
-`;
 
-interface inputProps {
-  hasText: boolean;
-}
-const Input = styled.textarea<inputProps>`
-  width: 100%;
-  height: 5.25rem;
-  border-radius: 0rem 0.625rem 0.625rem 0.625rem;
-  border: 1px solid #3a3a3a;
-  background: #fff;
-  padding: 0.56rem;
-  margin-top: 1rem;
-  color: ${(props) =>
-    props.hasText ? theme.colors.text : theme.colors.mutedText};
-  font-family: Pretendard;
-  font-size: 0.625rem;
-  font-weight: 300;
-  resize: none;
-  position: relative;
-`;
-const CharCount = styled.div`
-  position: absolute;
-  bottom: 0.5rem;
-  right: 0.56rem;
-  color: ${theme.colors.text};
-  font-size: 0.375rem;
-  font-style: normal;
-  font-weight: 300;
-`;
 export default function ToDeveloper() {
-  const [text, setText] = useState("");
-
+  const [inputs, setInputs] = useState([{ text: "" }]);
+  const addInputContainer = () => {
+    setInputs([...inputs, { text: "" }]);
+  };
   return (
     <Wrapper>
       <SettingHeader />
@@ -94,25 +56,9 @@ export default function ToDeveloper() {
           <Title>
             <Highlight>이런 기능</Highlight>있었으면 좋겠어요!
           </Title>
-          <BiMessageSquareEdit fontSize="2rem" />
+          <BiMessageSquareEdit fontSize="2rem" onClick={addInputContainer} />
         </TitleContent>
-
-        <InputContainer>
-          <Icon>
-            {text.length > 0 ? (
-              <FaCheckCircle color="#4CAF50" />
-            ) : (
-              <FaPlusCircle />
-            )}
-          </Icon>
-          <Input
-            value={text}
-            onChange={(e) => setText(e.target.value.slice(0, 100))}
-            hasText={text.length > 0}
-            placeholder="개발자에게 원하는 점을 적어주세요"
-          />
-          <CharCount>{text.length}/100</CharCount>
-        </InputContainer>
+        <SettingInput />
       </Container>
     </Wrapper>
   );
